requirements.txt
fastapi~=0.112.1
pydantic~=2.8.2
uvicorn~=0.30.6
python-dateutil~=2.9.0.post0